Refactor out two more methods
