+++ packages/client/src/Components/Mapping/helpers/updateRouteLineForForce.js
@@ -1,0 +1,8 @@
+import colorFor from './colorFor'
+
+updateRouteLineForForce (/* string */ force, /* Line */ line) {
+  const hisColor = colorFor(force)
+  line.setStyle({
+    color: hisColor
+  })
+}
